+++ twcore/src/twcore/core/game/Projectile.java
@@ -77,9 +77,9 @@ public class Projectile {
 		return dist;
 	}
 
-	public double getXLocation() { return x + vX * getAge() / (double) 10000.0; }
+	public double getXLocation() { return x + vX * getAge() / 10000.0; }
 
-	public double getYLocation() { return y + vY * getAge() / (double) 10000.0; }
+	public double getYLocation() { return y + vY * getAge() / 10000.0; }
 
 	public int getXVelocity() { return vX; }
